backend/requirements.txt
# Backend dependencies for LiveKit voice assistant (recommended balanced versions)
# Target Python: 3.11+ recommended

# Web
fastapi==0.95.2
uvicorn[standard]==0.22.0

# LiveKit Python SDK
livekit==1.2.13

# STT / TTS / LLM helpers
faster-whisper>=0.7.0
TTS==0.22.0
transformers>=4.35.0

# Env + http client
python-dotenv==1.0.0
requests==2.31.0

# Numeric / audio
numpy==1.26.0
scipy==1.11.0
soundfile==0.12.1
